requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
openai==1.5.0
python-dotenv==1.0.0
python-telegram-bot==20.7
